requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
paddlepaddle==2.6.1
paddleocr==2.7.0
pillow==10.1.0
python-multipart==0.0.6
